+++ lib/ok_image.dart
@@ -1,4 +1,4 @@
 library ok_image;
 
 export 'src/ok_image.dart';
-export 'src/cache_delegate.dart';
+export 'src/cache/cache_delegate.dart';
