+++ lib/src/cache/cache_delegate.dart
