+++ lib/src/cache/default_cache_delegate.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 
-import 'package:ok_image/ok_image.dart';
+import 'package:ok_image/src/cache/cache_delegate.dart';
 import 'package:ok_image/src/cache/cache_manager.dart';
 
 Future<Uint8List> defaultCache(String url, DefaultFuture createDefaultFuture, {bool followRedirects}) async {
